+++ CHANGELOG.rst
@@ -2,6 +2,14 @@
 Changelog for package pylon_ros2_camera
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+2.0.4 (2024-02-12)
+-------------------
+* Fix access to chunk data: line status all, counter value, frame counter, exposure, timestamp. Adding dedicated test script.
+
+2.0.3 (2023-12-01)
+-------------------
+* Fix PylonROS2CameraNode plugin-component mismatch 
+
 2.0.2 (2023-11-01)
 -------------------
 * All streamable parameters are set, regardless of the loaded user set